# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Banque API",
        "description": "API de gestion bancaire avec comptes et transactions",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://bankmanager-api.onrender.com/api/v1",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Lister les transactions",
                "description": "Lister les transactions avec filtrage et pagination\n\nRécupère la liste paginée des transactions avec possibilité de filtrage",
                "operationId": "41f5865254fe3acd0cc877cdead327c1",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer par type",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "depot",
                                "retrait",
                                "transfert",
                                "virement"
                            ]
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "en_attente",
                                "validee",
                                "rejete"
                            ]
                        }
                    },
                    {
                        "name": "compte_id",
                        "in": "query",
                        "description": "Filtrer par compte (source ou destination)",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "date_debut",
                        "in": "query",
                        "description": "Date de début (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_fin",
                        "in": "query",
                        "description": "Date de fin (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid"
                                                    },
                                                    "numero": {
                                                        "type": "string"
                                                    },
                                                    "type": {
                                                        "type": "string"
                                                    },
                                                    "montant": {
                                                        "type": "number"
                                                    },
                                                    "devise": {
                                                        "type": "string"
                                                    },
                                                    "statut": {
                                                        "type": "string"
                                                    },
                                                    "date_transaction": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Créer une transaction",
                "description": "Créer une nouvelle transaction\n\nCrée une transaction bancaire avec validation métier",
                "operationId": "efa8c65e0b71d4cea8ee038d20d1063a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "montant"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "depot",
                                            "retrait",
                                            "transfert",
                                            "virement"
                                        ],
                                        "example": "transfert"
                                    },
                                    "montant": {
                                        "type": "number",
                                        "minimum": 100,
                                        "example": 50000
                                    },
                                    "devise": {
                                        "type": "string",
                                        "default": "XOF",
                                        "example": "XOF"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Paiement facture"
                                    },
                                    "compte_source_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "compte_destination_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "date_transaction": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-28"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction créée avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "numero": {
                                                    "type": "string"
                                                },
                                                "type": {
                                                    "type": "string"
                                                },
                                                "montant": {
                                                    "type": "number"
                                                },
                                                "statut": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "400": {
                        "description": "Erreur métier"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transactions/{transaction}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Récupérer une transaction",
                "description": "Récupérer une transaction spécifique",
                "operationId": "09f6470ee2e38b19800723bb389b4e1f",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "description": "UUID de la transaction",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la transaction",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "numero": {
                                                    "type": "string"
                                                },
                                                "type": {
                                                    "type": "string"
                                                },
                                                "montant": {
                                                    "type": "number"
                                                },
                                                "statut": {
                                                    "type": "string"
                                                },
                                                "compte_source": {
                                                    "type": "object"
                                                },
                                                "compte_destination": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}